fix problems with letrec-based polymorphism
The way that polymorphic recursion was implemented allowed a type
variable from the type of a non-polymorphic binding to be turned into
a for-all-bound type variable in another binding.

The solution is to mark such variables as ineligible for
generalization to for-all binding later. However, "later" should only
last as long as generalization of the immediately enclosing set of
bindings. The binding's contribution to an enclosing context's type
can turn the variable into a for-all-bound variable after all.

This repair partly goes back to the changes in a8ecbe1f6b.  The
timestamps that a8ecbe1f6b got rid of were supposed to do the job for
nesting (I think!), but timestamps do not have the right structure.
This new approach uses lists of symbols, which provide enough
information to reflect the tree structure of nested bindings.

This is a big change, and there's a significant chance of the type
checker now rejecting programs that it should accept (while it
certainly rejects some programs that it shouldn't accept).