avoid loading contracts when importing as typed
The expansion of a `plai-typed` module now puts all the code in a
`typed-plai` submodule, puts contract-wrapped exports in a
more nested `contract-wrapped` sub-submodule, and arranges for
the enclosing module to re-export `contract-wrapped`. Importing
the module from another typed context uses the `plai-typed`
submodule directly, bypassing (in the sense of not even loading)
contracts.

Avoiding contract creation is not likely to matter for teaching
purposes, but this separation demonstrates an idea.